+++ app/admin/possessions.rb
@@ -1,5 +1,5 @@
 ActiveAdmin.register Possession do
-  permit_params :user_id, :product_id, :product_variant_id, :custom_product_id, :product_option_id, :prev_owned, :period_from, :period_to, :price_purchase, :price_purchase_currency, :price_sale, :price_sale_currency
+  permit_params :user_id, :product_id, :product_variant_id, :custom_product_id, :product_option_id, :prev_owned, :period_from, :period_to, :price_purchase, :price_purchase_currency, :price_sale, :price_sale_currency, :purchase_condition
 
   config.filters = false
 
@@ -35,6 +35,9 @@ ActiveAdmin.register Possession do
     end
     column "Sale", sortable: :price_sale do |entity|
       "#{entity.price_sale} #{entity.price_sale_currency}"
+    end
+    column :purchase_condition do |entity|
+      entity.purchase_condition_label || '—'
     end
     column :prev_owned
     actions