+++ app/controllers/custom_products_controller.rb
@@ -22,7 +22,10 @@ class CustomProductsController < ApplicationController
 
     possession = @user.possessions.find_by(custom_product_id: id)
     @possession = CustomProductPossessionPresenter.new(possession) if possession
-    @render_possession = @possession && (@possession.period_from || @possession.period_to || @possession.setup)
+    @render_possession_details = @possession && (
+      @possession.period_from || @possession.period_to || @possession.setup ||
+      @possession.price_purchase || @possession.price_sale || @possession.purchase_condition
+    )
 
     @custom_product = CustomProductPresenter.new(@user.custom_products.find(id))