+++ app/presenters/custom_product_possession_presenter.rb
@@ -10,6 +10,7 @@ class CustomProductPossessionPresenter < CustomProductPresenter
   delegate :price_purchase_currency, to: :@object
   delegate :price_sale, to: :@object
   delegate :price_sale_currency, to: :@object
+  delegate :purchase_condition, to: :@object
   delegate :images, to: :@custom_product
 
   def initialize(object)