+++ db/migrate/20260512120000_add_purchase_condition_to_possessions.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddPurchaseConditionToPossessions < ActiveRecord::Migration[8.1]
+  def change
+    add_column :possessions, :purchase_condition, :integer, null: true
+  end
+end